+++ CHANGELOG.md
@@ -6,11 +6,15 @@
 ### Features
 1. New Properties Sheet format added for 
     1. Block Storage Volume
+2. Add Option to Generate Terraform on Save
+3. Store Save Filename within model to allow for it to be populated during Save.
 ### Bug Fixes
 1. Move clone functionality to the top level class to resolve duplicate resource_name issue.
 2. Resolve issue where vcn cidr_blocks were being converted from an array to a string.
 3. Resolve Delete issue which caused the subnet.js to fail to load on FireFox.
 4. Resolve issue where exporting to Resource Manager never retrieved the Regions / Compartments
+5. Fix issue where Load Balancers with NSGs were generating invalid Terraform
+6. Resolve issue where Region List was being returned as undefined.
 
 
 ## Version 0.33.0